fixed floating issue

+++ PythonPackage/superphot/utils/explore_prf.py
@@ -1317,8 +1317,8 @@ def extract_pixel_data(cmdline_args, image_slices, sources=None):
         if sources is None:
             #pylint: enable=no-member
             sources = get_source_info(
-                pixel_array=frame[first_hdu].data,
-                stddev_array=frame[first_hdu + 1].data,
+                pixel_array=frame[first_hdu].data.astype(float),
+                stddev_array=frame[first_hdu + 1].data.astype(float),
                 mask_array=frame[first_hdu + 2].data.astype(c_char),
                 source_positions=get_source_positions(cmdline_args.catalogue,
                                                       trans_fname,
